# [POC] Add dogstatsd metric cache