### What does this PR do?

Add LRU cache to reduce time spent by `dogstatsd` in parsing metrics. It also reduce the allocation.

With LRU cache:

- 21% reduction in time per op
- 25% reduction in allocs/op

```
# Benchmark before (master branch)
BenchmarkParsePacket1-8      2000000           873 ns/op         431 B/op         12 allocs/op
BenchmarkParsePacket2-8      2000000           825 ns/op         425 B/op         12 allocs/op
BenchmarkParsePacket3-8      2000000           837 ns/op         445 B/op         12 allocs/op

Average ns/op: 2535
Average allocs/op: 12


# Benchmark with cache (this PR)
BenchmarkParsePacketWithCache1-8     2000000           663 ns/op         397 B/op          9 allocs/op
BenchmarkParsePacketWithCache2-8     2000000           646 ns/op         416 B/op          9 allocs/op
BenchmarkParsePacketWithCache3-8     2000000           687 ns/op         429 B/op          9 allocs/op

Average ns/op: 1996
Average allocs/op: 9


(1996-2535)/2535 = -0,2126
```

## Further optimisation: Share MetricSample fields

We can make `MetricCacheItem` a subset of `metrics.MetricSample`, but that will lead to more changes across the code base since `metrics.MetricSample` is widely used.

`MetricCacheItem` can be cached (LRU) and shared for all metrics with same characteristics.

Probably 10-20% reduction in ns/op and allocs/op.

```
type MetricSample struct {
	CacheSample *MetricCacheItem
	Value      float64
	RawValue   string
	Timestamp  float64
}

type MetricCacheItem struct {
	Name       string
	Mtype      metrics.MetricType
	Tags       []string
	Host       string
	SampleRate float64
}
```

Note:
- There is probably a better name for `MetricCacheItem`
- MetricCacheItem is a bit different than the concept MetricSampleContext. SampleRate and Mtype are also in MetricCacheItem.


### Motivation

Performance & Memory.

### Additional Notes